requirements.txt
streamlit>=1.20.0
langchain>=0.0.250
langchain_community>=0.0.1
pinecone-client>=2.0.0
openai>=0.27.0
google-cloud-dialogflow

